feat: seprate email meta component
